Clean dynamical two-compartment model, dirty notebook
Clean implementation of two compartment neuron model with somatic output
fed to dendrite and vice-versa. Input from other compartments gets added
to external input before filtering and applying a nonlinearity.

Notebook is dirty because plots are a work in progress!

Eventually TwoCompartmentNeuron and related classes should be extracted
from this notebook and put in an installable package for others to use.