Finish refactoring base SNN
Finish extracting components of Peter's spiking neural network (SNN) to
training/model_components.py and update
training/spytorch-tutorial-1.ipynb to reflect this change. This model
takes input only to the dendritic compartment, which has a recurrent
connection with the soma. There are no parallel subunits in the
dendritic compartment as of this commit.