Dirty debugging
The changes in this commit were over a week old at the time of the
commit.

Tweaks to library components to try to figure out why some models were
not being trained in tutorial 1 notebook. They should probably be
reverted because they turned out not to be the cause of the problem.
(The cause was related to lambda functions being stored in a dict in the
notebook itself.)